+++ js/main.js
@@ -1,7 +1,7 @@
 import {generatePhotos} from './mocks/photos.js';
 import {generateThumbnails} from './thumbnails.js';
 import {generateBigPicturesOpen} from './big-pictures.js';
-import './img-upload-form.js';
+import './img-upload-form/show-form.js';
 
 const MAX_PHOTO_COUNT = 25;
 const thumbnailsContainer = document.querySelector('.pictures');
